# routes/product.js
const router = require('express').Router()
const Product = require('../models/Product')
const { verifyToken, verifyTokenAndAuth, verifyTokenAndAdmin } = require('./verifyToken')

// CREATE
router.post('/create', verifyTokenAndAdmin, async (req, res) => {
    const newProduct = new Product(req.body)

    try {
        const savedProduct = await newProduct.save()
        res.status(200).json(savedProduct)
    } catch (error) {
        res.status(500).json(error)
    }
})

// UPDATE
router.put("/:id", verifyTokenAndAdmin, async (req, res) => {
    try {
        const updatedProduct = await Product.findByIdAndUpdate(req.params.id, {
            $set: req.body
        }, { new: true })
        res.status(200).json(updatedProduct)
    } catch (error) {
        res.status(500).json(error)
    }
})

// DELETE
router.delete('/:id', verifyTokenAndAdmin, async (req, res) => {
    try {
        await Product.findByIdAndDelete(req.params.id)
        res.status(200).json('Deleted Product!')
    } catch (error) {
        res.status(500).json(error)
    }
})

// GET PRODUCT BY ID
router.get('/find/:id', async (req, res) => {
    try {
        const product = await Product.findById(req.params.id)
        res.status(200).json(product)
    } catch (error) {
        res.status(500).json(error)
    }
})

// GET PRODUCT BY ID
router.get('/', async (req, res) => {
    const qNew = req.query.new
    try {
        let products

        if (qNew) {
            products = await Product.find().sort({ createdAt: -1 }).limit(5)
        } else {
            products = await Product.find()
        }
        res.status(200).json(products)
    } catch (error) {
        res.status(500).json(error)
    }
})

module.exports = router